Seperating functions into classes
